+++ web/src/features/moreCast2/components/validatedForecastCell.test.tsx
@@ -3,7 +3,7 @@ import { GridRenderCellParams } from '@mui/x-data-grid-pro'
 import ValidatedForecastCell from '@/features/moreCast2/components/ValidatedForecastCell'
 import { initialState } from '@/features/moreCast2/slices/validInputSlice'
 import { Provider } from 'react-redux'
-import { buildTestStore } from '@/features/moreCast2/components/testHelper.test'
+import { buildTestStore } from '@/features/moreCast2/components/testHelper'
 
 const params: Pick<GridRenderCellParams, 'row' | 'formattedValue'> = {
   row: undefined,
